Move test code to root directory
